Doc update + move around media embedding service
